+++ lib/src/backend/configuration.dart
@@ -7,23 +7,25 @@ import 'package:yaml/yaml.dart';
 class Configuration {
   bool captureScreenshotOnError = false;
   bool captureScreenshotOnSuccess = false;
+  Map<String, dynamic> capabilities = {};
 
   /// Relative path to the working directory in which the tests are executed.
   String screenshotPath = 'test_screenshots';
 
-  Configuration._(Map<String, dynamic> v) {
+  Configuration.fromMap(Map<String, dynamic> v) {
     var screen = v['screenshot'] ?? {};
     captureScreenshotOnError = screen['onError'] ?? false;
     captureScreenshotOnSuccess = screen['onSuccess'] ?? false;
     screenshotPath = screen['output'] ?? 'test_screenshots';
+    capabilities = v['capabilities'] ?? {};
   }
 }
 
 /// Loads the configuration from the working directory.
 Future<Configuration> loadConfig() async {
   var file = new File('dart_webdriver.yaml');
-  if (!await file.exists()) return new Configuration._({});
+  if (!await file.exists()) return new Configuration.fromMap({});
 
   var yaml = loadYaml(await file.readAsString());
-  return new Configuration._(yaml);
+  return new Configuration.fromMap(yaml);
 }