+++ lib/test_webdriver.dart
@@ -5,6 +5,7 @@ import 'dart:mirrors';
 
 import 'package:test/test.dart' as ts;
 import 'package:webdriver/io.dart';
+import 'package:webdriver/support/async.dart';
 import 'src/backend/suite.dart';
 
 export 'package:test/test.dart';
@@ -42,14 +43,38 @@ Function withDriver(Function fn) {
 /// tries to load it using the loader of the current suite and injects it into
 /// the function. The test case performed in this function is wrapped and
 /// the suite will perform followup functionality like taking screenshots.
-Function withPO(Function fn, {String screenshotName}) {
+///
+/// Set [useWaitFor] to false in case the pageloader should instantly throw
+/// an exception if a pageobject is not present on test execution. Otherwise the
+/// pageloader will try to load the pageobject until the test times out.
+///
+/// Use [timeout] to define a separate timeout for the pageloader to fetch the
+/// pageobject apart from the test specific timeout. For further details on the
+/// default timeout, see [test] package documentation.
+Function withPO(Function fn,
+    {String screenshotName,
+    bool useWaitFor: true,
+    Duration timeout: const Duration(days: 1)}) {
   var suite = Suite.current;
   var mirr = reflect(fn) as ClosureMirror;
 
+  assert(mirr.function.parameters.isNotEmpty,
+      'expect withPO body to have an argument');
+
+  var argType = mirr.function.parameters.first.type.reflectedType;
+
   return () async {
     return suite.run(() async {
-      var po = await suite.loader
-          .getInstance(mirr.function.parameters.first.type.reflectedType);
+      var po;
+
+      if (useWaitFor) {
+        po = await waitFor(() => suite.loader.getInstance(argType),
+            matcher: ts.isNot(
+                ts.throwsA(new ts.isInstanceOf<NoSuchElementException>())),
+            timeout: timeout);
+      } else {
+        po = await suite.loader.getInstance(argType);
+      }
 
       var exception;
       try {