+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# 0.0.1
+## 0.0.2
 
-- Initial release
+* Support waitFor functionality within `withPO` 
+* Support capabilities configuration within configuration. Includes optional configuration of headless mode instead of explicit set (see README.md)
+* Provide timeout settings
+
+## 0.0.1
+
+* Initial release